# [WIP] Track folder changes to update USN journal