--- a/GVFS/GVFS.FunctionalTests.Windows/GVFS.FunctionalTests.Windows.csproj
+++ b/GVFS/GVFS.FunctionalTests.Windows/GVFS.FunctionalTests.Windows.csproj
@@ -97,6 +97,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Windows\Tests\WindowsDiskLayoutUpgradeTests.cs" />
     <Compile Include="Windows\Tests\JunctionAndSubstTests.cs" />
+    <Compile Include="Windows\Tests\WindowsFolderUsnUpdate.cs" />
     <Compile Include="Windows\Tests\WindowsTombstoneTests.cs" />
     <Compile Include="Windows\Tests\WindowsUpdatePlaceholderTests.cs" />
     <Compile Include="Windows\Tests\WindowsFileSystemTests.cs" />