BuildXL was having issues detecting that a folder had changes when the projection changed because VFSForGit was simply updating its in memory projection and not doing anything with the file system that would trigger and update to the USN. It would hand back the new items in the directory when projfs would ask for the items for enumeration the next time an enumeration happened.

This change is to make the folder get a new USN when VFSForGit determines that a folder has changed.  This is done by keeping a SHA1 hash of all the SHAs of the files that are in the projection and when the projection changes calculate the new SHA for the folder and compare with what it had before.  If it changed then write and delete a file in the folder to trigger the USN update.

Something to note. We don't need any upgrade step because we are already storing the SHA for the folder in the placeholder database but it is currently `null` for all folders so the first time there is a projection change it will cause all the folders to get the calculated SHA for the folder and update the placeholder database. 

Todos:
- [x] Error handling
- [x] Functional tests
- [x] Check Memory and Performance
- [x] Give version to BuildXL to test with