@kewillford 2020-04-22
I didn't find any change in the memory footprint with this version.

Differences in performance:
|Command|Folder Placeholders|Time before|Time After|
|---|---:|---:|---:|
|Repo 1 `checkout`|~100K|12s|13s|
|Repo 2 `checkout`|~150K|8.3s|9.5s|

So there is ~10% perf regression in `checkout` when there a greater than ~100K folder placeholders.  With fewer than 10K placeholders there was not a significant increase in the timings.