@benwitman 2020-05-15
We don’t want to hash the shas, we only want to hash the names.  That should be changed for the file case too