@kewillford 2020-05-15
> We don’t want to hash the shas, we only want to hash the names. That should be changed for the file case too

If there is a pure virtual file for the folder and the file's name doesn't change but the content does, the USN will not be updated for the folder.  Will that be okay because all the files of a folder will at least have a placeholder on disk?