@narasamdya 2020-05-16
I mentioned in the other comment that having the shas will not cause overbuild (I'll create a unit test on BuildXL's side). But it forces BuildXL to perform extra validation, that could've been avoided.

---
In reply to: [629356194](https://github.com/microsoft/VFSForGit/pull/1648#issuecomment-629356194) [](ancestors = 629356194)